Prevent validators from getting inserted twice
Related to this flask-admin issue: https://github.com/flask-admin/flask-admin/issues/1094

If you override ```field_args``` with your own dict, the changes made to ```kwargs``` in ```convert()``` will also be made to the dict you used to override ```field_args```. The objects that get copied to ```kwargs``` from ```field_args``` are only references, so you need to deepcopy to prevent the issue.